LayoutTests/http/tests/site-isolation/device-media-features-in-cross-origin-iframe.html

<!DOCTYPE html><!-- webkit-test-runner [ SiteIsolationEnabled=true ] -->
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<iframe id="iframeTest" src="http://localhost:8000/site-isolation/resources/device-media-features-iframe.html"></iframe>
<script>
description("device-width, device-height, and device-aspect-ratio media features should match screen dimensions inside a cross-origin iframe under site isolation.");

var jsTestIsAsync = true;
var result = null;

window.addEventListener("message", function(event) {
    result = event.data;
    shouldBeTrue("result.screenWidth > 0");
    shouldBeTrue("result.screenHeight > 0");
    shouldBeFalse("result.deviceWidthZeroMatches");
    shouldBeFalse("result.deviceHeightZeroMatches");
    shouldBeTrue("result.minDeviceWidthMatchesScreen");
    shouldBeTrue("result.maxDeviceWidthMatchesScreen");
    shouldBeTrue("result.minDeviceHeightMatchesScreen");
    shouldBeTrue("result.maxDeviceHeightMatchesScreen");
    shouldBeTrue("result.deviceAspectRatioMatchesScreen");
    finishJSTest();
}, false);
</script>
</body>
</html>
