LayoutTests/http/tests/site-isolation/device-media-features-in-cross-origin-iframe-expected.txt

device-width, device-height, and device-aspect-ratio media features should match screen dimensions inside a cross-origin iframe under site isolation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS result.screenWidth > 0 is true
PASS result.screenHeight > 0 is true
PASS result.deviceWidthZeroMatches is false
PASS result.deviceHeightZeroMatches is false
PASS result.minDeviceWidthMatchesScreen is true
PASS result.maxDeviceWidthMatchesScreen is true
PASS result.minDeviceHeightMatchesScreen is true
PASS result.maxDeviceHeightMatchesScreen is true
PASS result.deviceAspectRatioMatchesScreen is true
PASS successfullyParsed is true

TEST COMPLETE
